Mixins & Generic API View -> Refactor category and product views to use class-based views, enhancing structure and supporting additional HTTP methods for better management.

+++ product/views.py
@@ -3,84 +3,39 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
+from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
 
 from product.models import Product, Category
 from product.serializer import Product_Serializer, Category_Serializer
 
 from django.db.models import Count
 
-
-class View_Products( APIView ):
-    def get(self, request):
-        products = Product.objects.select_related('category').all()
-        serializer = Product_Serializer(products, many=True, context={'request': request})
-        return Response(serializer.data)
+class Product_List( ListCreateAPIView ):
+    queryset = Product.objects.select_related('category').all()
+    serializer_class = Product_Serializer
     
-    def post(self, request):
-        serializer = Product_Serializer(data=request.data, context={'request': request})  # Deserializer
-        serializer.is_valid(raise_exception=True)
-
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+    def get_serializer_context(self):
+        return {'request': self.request}
     
 
-class View_Specific_Product( APIView ):
+class Product_Details( RetrieveUpdateDestroyAPIView ):
+    queryset = Product.objects.all()
+    serializer_class = Product_Serializer
 
-    def get( self, request, pk ):
-        product = get_object_or_404( Product, pk = pk )
-        serializer = Product_Serializer( product, context={'request': request} )
-        return Response( serializer.data )
-    
-    def put( self, request, pk ):
-        product = get_object_or_404( Product, pk = pk )
-        serializer = Product_Serializer(product, data=request.data, context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response( serializer.data )
-    
     def delete(self, request, pk):
-        product = get_object_or_404( Product, pk = pk )
+        product = get_object_or_404(Product, pk=pk)
+        if product.stock > 0:
+            return Response({'message': 'You can\'t delete an available product'})  
+        
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-        
 
-class View_Categories( APIView ):
-    def get(self, request):
-        categories = Category.objects.annotate(product_count=Count('products'))
-        serializer = Category_Serializer( categories, many=True )
 
-        return Response( serializer.data )
+class Category_List( ListCreateAPIView ):
+    queryset = Category.objects.annotate(product_count=Count('products'))
+    serializer_class = Category_Serializer
     
-    def post(self, request):
-        serializer = Category_Serializer(data=request.data)     # Deserializer
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-
-class View_Specific_Category( APIView ):
-    def get(self, request, pk):
-        category = get_object_or_404(
-            Category.objects.annotate(product_count=Count('products')), 
-            pk = pk 
-        )
-        serializer = Category_Serializer( category )
-
-        return Response( serializer.data )
-    
-    def put(self, request, pk):
-        category = get_object_or_404(
-            Category.objects.annotate(product_count=Count('products')), 
-            pk = pk 
-        )
-        serializer = Category_Serializer(category, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data)
-    
-    def delete(self, request, pk):
-        category = get_object_or_404( Category, pk = pk )
-        category.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+class Category_Details( RetrieveUpdateDestroyAPIView ):
+    queryset = Category.objects.annotate(product_count=Count('products'))
+    serializer_class = Category_Serializer
